+++ src/libzetscript/core/CScriptFunctionObject.cpp
@@ -34,7 +34,7 @@ tInfoVariableSymbol * CScriptFunctionObject::newVariableSymbol(int idxFunction){
 		tInfoVariableSymbol irs;
 		CScriptFunctionObject *irc = vec_script_function_object_node->at(idxFunction);
 
-		irs.idxSymbol = irc->object_info.local_symbols.m_registeredVariable.size();
+		irs.idxSymbol = (short)irc->object_info.local_symbols.m_registeredVariable.size();
 		irc->object_info.local_symbols.m_registeredVariable.push_back(irs);
 		return &irc->object_info.local_symbols.m_registeredVariable[irc->object_info.local_symbols.m_registeredVariable.size()-1];
 	}
